requirements.txt
asgiref==3.6.0
attrs==22.2.0
black==23.3.0
certifi==2022.12.7
charset-normalizer==2.0.12
click==8.1.3
Django==3.2
djangorestframework==3.12.4
flake8==6.0.0
idna==3.4
iniconfig==2.0.0
isort==5.12.0
mccabe==0.7.0
mypy-extensions==1.0.0
packaging==23.0
pathspec==0.11.1
platformdirs==3.2.0
pluggy==0.13.1
py==1.11.0
pycodestyle==2.10.0
pyflakes==3.0.1
PyJWT==2.1.0
pytest==6.2.4
pytest-django==4.4.0
pytest-pythonpath==0.7.3
pytz==2023.2
requests==2.26.0
sqlparse==0.4.3
toml==0.10.2
tomli==2.0.1
urllib3==1.26.15
